Expose ponyfill
